added inquire
