+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "Scale_Generator"
+name = "scale-generator"
 authors = ["Benjamin Goodheart"]
-version = "0.1.0"
+version = "0.1.1"
 edition = "2024"
 
 [dependencies]
@@ -10,3 +10,5 @@ rand = "0.9.2"
 rustyline = "17.0.2"
 cargo-wix = "0.3.9"
 clap = {version="4.6.0", features=["derive", "cargo"]}
+shlex = "1.3.0"
+inquire = "0.9.4"
